EA-4014 few tweaks
