+++ iiif-presentation-web/src/main/java/eu/europeana/api/iiif/utils/IIIFConstants.java
@@ -15,7 +15,10 @@ public class IIIFConstants {
     public static final String BEAN_USER_SET_API_CLIENT = "userSetApiClient";
     public static final String BEAN_COLLECTION_V2_GENERATOR = "collectionV2Generator";
     public static final String BEAN_COLLECTION_V3_GENERATOR = "collectionV3Generator";
-    public static final String BEAN_JSON_MAPPER = "jsonMapper";
+    public static final String BEAN_IIIF_JSON_HANDLER = "iifJsonHandler";
+
+    public static final String BEAN_V2_JSON_MAPPER = "v2JsonMapper";
+    public static final String BEAN_V3_JSON_MAPPER = "v3JsonMapper";
 
 
     public static final String ACCEPT  = "Accept=";
